# Fix `format_command`'s terrible performance