@casperisfine 2022-11-25
I updated the implementation to generate the same output as before based on the test suite, so it's not longer as fast but still way better:

```
Warming up --------------------------------------
            original     4.000  i/100ms
                 opt    63.244k i/100ms
Calculating -------------------------------------
            original     43.899  (± 2.3%) i/s -    220.000  in   5.012839s
                 opt    627.939k (± 0.8%) i/s -      3.162M in   5.036167s

Comparison:
            original:       43.9 i/s
                 opt:   627939.2 i/s - 14304.28x  (± 0.00) faster
```