@benoittgt 2022-11-25
Shouldn't we have a non regression performance test ? I know it is a pain to maintain, but I am wondering if could help to avoid futur regression.